refactor: split image_gen.c into compute units to meet 1000 line limit
